+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.1] - 2026-01-19
+
+### Fixed
+- Fixed All Access pass handling to work with array format returned by `edd_all_access_get_customer_passes()`
+- Passes are now identified by `payment_id` instead of `id` since the array format lacks an `id` field
+
 ## [1.2.0] - 2026-01-19
 
 ### Changed